requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Azure Cloud Integration
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Web Application Framework
streamlit>=1.29.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pickle-mixin>=1.0.2
python-dateutil>=2.8.2
